pyproject.toml
[project]
name = "nifty_anilist"
version = "0.1.5"
description = "Utilities for the Anilist GraphQL API."
authors = [
    { name="Alexey Polishchuk", email="alexpolishchukmail@gmail.com" }
]
readme = "README.md"
license = "MIT"
license-files = ["LICENSE"]
requires-python = ">=3.12"
dependencies = [
    "asyncio>=4.0.0",
    "colorama>=0.4.6",
    "httpx>=0.28.1",
    "keyring>=25.6.0",
    "pydantic>=2.11.9",
    "pydantic-settings>=2.11.0",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.1.1",
    "selenium>=4.35.0",
    "structlog>=25.4.0",
]

[project.urls]
"Repository" = "https://github.com/AlexPolGit/nifty-anilist"

[tool.setuptools]
include-package-data = true

[tool.setuptools.package-data]
"nifty_anilist" = ["*.graphql"]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[dependency-groups]
dev = [
    "ariadne-codegen>=0.15.3",
    "pre-commit>=4.3.0",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
]

[tool.ariadne-codegen]
enable_custom_operations = true
target_package_name = "client"
target_package_path = "nifty_anilist"
remote_schema_url = "https://graphql.anilist.co"
